Move runtime_mode and agas_service_mode to runtime_configuration module
