+++ docs/CMakeLists.txt
@@ -29,7 +29,6 @@ set(doxygen_dependencies
     "${PROJECT_SOURCE_DIR}/hpx/runtime/get_worker_thread_num.hpp"
     "${PROJECT_SOURCE_DIR}/hpx/runtime/launch_policy.hpp"
     "${PROJECT_SOURCE_DIR}/hpx/runtime/report_error.hpp"
-    "${PROJECT_SOURCE_DIR}/hpx/runtime/runtime_mode.hpp"
     "${PROJECT_SOURCE_DIR}/hpx/runtime/set_parcel_write_handler.hpp"
     "${PROJECT_SOURCE_DIR}/hpx/runtime/shutdown_function.hpp"
     "${PROJECT_SOURCE_DIR}/hpx/runtime/startup_function.hpp"
