+++ hpx/runtime_fwd.hpp
@@ -25,7 +25,7 @@
 #include <hpx/runtime/naming_fwd.hpp>
 #include <hpx/runtime/report_error.hpp>
 #include <hpx/runtime/runtime_fwd.hpp>
-#include <hpx/runtime/runtime_mode.hpp>
+#include <hpx/runtime_configuration/runtime_mode.hpp>
 #include <hpx/runtime/set_parcel_write_handler.hpp>
 #include <hpx/runtime/shutdown_function.hpp>
 #include <hpx/runtime/startup_function.hpp>
