+++ hpx/util/command_line_handling.hpp
@@ -10,7 +10,7 @@
 #include <hpx/config.hpp>
 #include <hpx/hpx_init.hpp>
 #include <hpx/program_options.hpp>
-#include <hpx/runtime/runtime_mode.hpp>
+#include <hpx/runtime_configuration/runtime_mode.hpp>
 #include <hpx/functional/function.hpp>
 #include <hpx/util/manage_config.hpp>
 #include <hpx/runtime_configuration/runtime_configuration.hpp>
