+++ libs/resource_partitioner/include/hpx/resource_partitioner/detail/create_partitioner.hpp
@@ -11,7 +11,7 @@
 #include <hpx/functional/bind_back.hpp>
 #include <hpx/functional/function.hpp>
 #include <hpx/resource_partitioner/partitioner_fwd.hpp>
-#include <hpx/runtime/runtime_mode.hpp>
+#include <hpx/runtime_configuration/runtime_mode.hpp>
 #include <hpx/runtime_configuration/find_prefix.hpp>
 
 #include <hpx/program_options.hpp>
