+++ libs/resource_partitioner/include/hpx/resource_partitioner/detail/partitioner.hpp
@@ -13,7 +13,7 @@
 #include <hpx/datastructures/tuple.hpp>
 #include <hpx/program_options.hpp>
 #include <hpx/resource_partitioner/partitioner.hpp>
-#include <hpx/runtime/runtime_mode.hpp>
+#include <hpx/runtime_configuration/runtime_mode.hpp>
 #include <hpx/synchronization/spinlock.hpp>
 #include <hpx/threading_base/scheduler_mode.hpp>
 #include <hpx/topology/cpu_mask.hpp>
