+++ libs/runtime_configuration/include/hpx/runtime_configuration/agas_service_mode.hpp
@@ -1,0 +1,22 @@
+//  Copyright (c) 2007-2014 Hartmut Kaiser
+//  Copyright (c) 2011      Bryce Lelbach
+//
+//  SPDX-License-Identifier: BSL-1.0
+//  Distributed under the Boost Software License, Version 1.0. (See accompanying
+//  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+
+#ifndef HPX_RUNTIME_CONFIGURATION_AGAS_SERVICE_MODE_HPP
+#define HPX_RUNTIME_CONFIGURATION_AGAS_SERVICE_MODE_HPP
+
+namespace hpx { namespace agas
+{
+    enum service_mode
+    {
+        service_mode_invalid = -1,
+        service_mode_bootstrap = 0,
+        service_mode_hosted = 1
+    };
+}}
+
+#endif /*HPX_RUNTIME_CONFIGURATION_AGAS_SERVICE_MODE_HPP*/
+
