+++ src/hpx_wrap.cpp
@@ -26,7 +26,7 @@ namespace hpx_start
 
 #include <hpx/hpx_finalize.hpp>
 #include <hpx/hpx_init.hpp>
-#include <hpx/runtime/runtime_mode.hpp>
+#include <hpx/runtime_configuration/runtime_mode.hpp>
 #include <hpx/functional/function.hpp>
 
 #include <string>
