+++ src/runtime.cpp
@@ -206,34 +206,6 @@ namespace hpx
     ///////////////////////////////////////////////////////////////////////////
     namespace strings
     {
-        char const* const runtime_mode_names[] =
-        {
-            "invalid",    // -1
-            "console",    // 0
-            "worker",     // 1
-            "connect",    // 2
-            "default",    // 3
-        };
-    }
-
-    char const* get_runtime_mode_name(runtime_mode state)
-    {
-        if (state < runtime_mode_invalid || state >= runtime_mode_last)
-            return "invalid (value out of bounds)";
-        return strings::runtime_mode_names[state+1];
-    }
-
-    runtime_mode get_runtime_mode_from_name(std::string const& mode)
-    {
-        for (std::size_t i = 0; i < runtime_mode_last; ++i) {
-            if (mode == strings::runtime_mode_names[i])
-                return static_cast<runtime_mode>(i-1);
-        }
-        return runtime_mode_invalid;
-    }
-
-    namespace strings
-    {
         char const* const runtime_state_names[] =
         {
             "state_invalid",      // -1
